+++ packages/ui-components/src/chat/chat.stories.tsx
@@ -17,6 +17,7 @@
 import { Meta } from '@storybook/react';
 import * as React from 'react';
 
+import { UserData } from '../types';
 import { default as ChatComponent } from './chat';
 
 export default {
@@ -24,32 +25,58 @@ export default {
     title: 'UI Components/Chat',
 } as Meta;
 
+const Hagrid: UserData = {
+    name: 'Rubeus Hagrid',
+    id: 'hagrid_id',
+};
+
+const Harry: UserData = {
+    name: 'Harry Potter',
+    id: 'harry_id',
+};
+
 const messages = [
     {
         id: 'yYZ1_TTZbL9is7RhG_C0l',
-        message:
-            'For uploading templates, we would need to provide a user the interface to select associated folders and files along with the dbook they wish to save to the registry.',
-        created_at: '2024-04-03T10:34:05.502Z',
-        updated_at: '2024-04-03T10:34:05.502Z',
+        message: '{puts cake down} \n Excuse me, who are you?',
+        created_at: '2024-04-03T10:34:05.167Z',
+        updated_at: '2024-04-03T10:34:05.167Z',
+        user: Harry,
     },
     {
         id: 'oGbwtu9PHMVNYbFfxN7Br',
-        message:
-            'Do we want to consider some of the management of metacell / template registries as a UI/UX flow occurring outside of Studio itself? i.e. it is a platform level user journey?',
+        message: "Rubeus Hagrid. Keeper of keys and grounds at Hogwarts. Course, you'll know all about Hogwarts.",
         created_at: '2024-04-03T10:34:17.167Z',
         updated_at: '2024-04-03T10:34:17.167Z',
+        user: Hagrid,
     },
     {
         id: 'ocRSEZU9DeILt8MvyKu0b',
-        message: 'This is another test message, which has been edited!',
+        message: 'Sorry, no.',
         created_at: '2024-04-03T10:34:26.944Z',
         updated_at: '2024-04-03T10:34:45.061Z',
+        user: Harry,
     },
     {
         id: 'ctDqA50c0b13FQKY0E1tm',
-        message: 'test\n**abc**\nhello there',
+        message: "No? Blimey, Harry, didn't you ever wonder where your mum and dad learned it all?",
         created_at: '2024-04-03T15:55:05.031Z',
         updated_at: '2024-04-03T15:55:05.031Z',
+        user: Hagrid,
+    },
+    {
+        id: 'ocRSEZU9DeILt8MvyKu07',
+        message: 'Learnt what?',
+        created_at: '2024-04-03T15:56:26.944Z',
+        updated_at: '2024-04-03T15:56:45.061Z',
+        user: Harry,
+    },
+    {
+        id: 'ctDqA50c0b13FQKY0E1tp',
+        message: "You're a **wizard**, Harry.",
+        created_at: '2024-04-03T15:57:26.944Z',
+        updated_at: '2024-04-03T15:57:26.944Z',
+        user: Hagrid,
     },
 ];
 
@@ -58,7 +85,7 @@ export const Chat = (): JSX.Element => {
 
     return (
         <div style={{ backgroundColor: 'black', display: 'flex' }}>
-            <ChatComponent onUpdate={setValue} value={value} />
+            <ChatComponent activeUser={Harry} onUpdate={setValue} value={value} />
         </div>
     );
 };