+++ packages/ui-components/src/chat/message.tsx
@@ -18,7 +18,7 @@ import { format, parseISO } from 'date-fns';
 import { isEqual } from 'lodash';
 import * as React from 'react';
 
-import styled from '@darajs/styled-components';
+import styled, { useTheme } from '@darajs/styled-components';
 import { PenToSquare, Trash } from '@darajs/ui-icons';
 
 import Button from '../button/button';
@@ -62,11 +62,12 @@ const MessageWrapper = styled.div`
 const MessageTop = styled.div`
     display: flex;
     justify-content: space-between;
-
     width: 100%;
-    height: 1.625rem;
+    font-size: 0.875rem;
+`;
 
-    font-size: 0.8rem;
+const MessageTimestamp = styled.span`
+    font-size: 0.75rem;
     color: ${(props) => props.theme.colors.grey5};
 `;
 
@@ -77,6 +78,7 @@ const MessageBody = styled.span`
 `;
 
 const EditedText = styled.span`
+    align-self: end;
     font-size: 0.8rem;
     color: ${(props) => props.theme.colors.grey4};
 `;
@@ -113,11 +115,33 @@ const EditButtons = styled.div`
     justify-content: flex-end;
 `;
 
+const UserInfoWrapper = styled.div`
+    display: flex;
+    gap: 0.5rem;
+    align-items: center;
+`;
+
+const AvatarIcon = styled.div`
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    width: 2rem;
+    height: 2rem;
+
+    font-weight: 700;
+    color: white;
+
+    border-radius: 50%;
+`;
+
 export interface MessageProps extends InteractiveComponentProps<Message> {
     /** An optional onChange handler for listening to changes in the input */
     onChange?: (value: Message, e?: React.SyntheticEvent<HTMLInputElement>) => void | Promise<void>;
     /** An optional event listener for complete events (enter presses) */
     onDelete?: (id: string) => void | Promise<void>;
+    /** An optional flag to determine if the message is editable */
+    isEditable?: boolean;
 }
 
 /**
@@ -125,6 +149,39 @@ export interface MessageProps extends InteractiveComponentProps<Message> {
  */
 export function getFormattedTimestamp(date: string): string {
     return format(parseISO(date), 'HH:mm dd/MM/yyyy');
+}
+
+/**
+ * A function to assign a color to user token depending on their name
+ */
+function selectColor(name: string, colors: string[]): string {
+    // Convert the name to lowercase for consistency
+    const lowerCaseName = name.toLowerCase();
+
+    // Calculate the sum of ASCII values of the characters in the name
+    let asciiSum = 0;
+    for (const char of lowerCaseName) {
+        asciiSum += char.charCodeAt(0);
+    }
+
+    // Use the remainder to select a color
+    const colorIndex = asciiSum % colors.length;
+    return colors[colorIndex];
+}
+
+/**
+ * A function to get the user's initials
+ */
+function getInitials(name: string): string {
+    const parts = name.trim().split(/\s+/);
+    let initials = parts[0][0];
+
+    // If there's a second part, add its first character, so we only get two letter initials
+    if (parts.length > 1) {
+        initials += parts[parts.length - 1][0];
+    }
+
+    return initials.toUpperCase();
 }
 
 /**
@@ -140,12 +197,24 @@ export function processText(text: string): string {
  * @param {MessageProps} props - the component props
  */
 function MessageComponent(props: MessageProps): JSX.Element {
+    const theme = useTheme();
     const [editMode, setEditMode] = React.useState(false);
     const [editMessage, setEditMessage] = React.useState(props.value.message);
     const [localMessage, setLocalMessage] = React.useState(props.value);
     if (props.value && !isEqual(props.value, localMessage)) {
         setLocalMessage(props.value);
     }
+
+    // List of colors for user token to pick from
+    const tokenColors = [
+        theme.colors.secondary,
+        theme.colors.violet,
+        theme.colors.turquoise,
+        theme.colors.purple,
+        theme.colors.teal,
+        theme.colors.orange,
+        theme.colors.plum,
+    ];
 
     const onAccept = (): void => {
         // if the message hasn't changed, just close the edit mode
@@ -176,15 +245,21 @@ function MessageComponent(props: MessageProps): JSX.Element {
     return (
         <MessageWrapper className={props.className} style={props.style}>
             <MessageTop>
-                <div>
-                    {getFormattedTimestamp(props.value.created_at)}
+                <UserInfoWrapper>
+                    <AvatarIcon style={{ backgroundColor: selectColor(localMessage.user.name, tokenColors) }}>
+                        {getInitials(localMessage.user.name)}
+                    </AvatarIcon>
+                    <div style={{ display: 'flex', flexDirection: 'column' }}>
+                        {localMessage.user.name}
+                        <MessageTimestamp>{getFormattedTimestamp(props.value.created_at)}</MessageTimestamp>
+                    </div>
                     {localMessage.updated_at !== localMessage.created_at && (
                         <Tooltip content={getFormattedTimestamp(props.value.updated_at)}>
                             <EditedText> (edited)</EditedText>
                         </Tooltip>
                     )}
-                </div>
-                {!editMode && (
+                </UserInfoWrapper>
+                {!editMode && props.isEditable && (
                     <InteractiveIcons>
                         <EditIcon data-testid="message-edit-button" onClick={() => setEditMode(true)} role="button" />
                         <DeleteIcon data-testid="message-delete-button" onClick={onDelete} role="button" />