+++ packages/ui-components/docs/changelog.md
@@ -8,6 +8,8 @@ title: Changelog
 -   `Chat` component now displays next line breaks.
 -   `Textarea` component when `onComplete` prop is passed adds next line with shift + enter. 
 -   Added a new `Markdown` component.
+-   **Breaking** `Chat` now takes a `active_user` prop, and display user info in each message.
+-   **Breaking** `Chat` component now only allows users that wrote a message to edit it.
 
 ## 1.8.0